tests/wpt/web-platform-tests/IndexedDB/idbfactory_open2.htm

<!DOCTYPE html>
<title>IDBFactory.open() - database 'name' and 'version' are correctly set</title>
<link rel="author" title="Microsoft" href="http://www.microsoft.com">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=support.js></script>

<script>
    var open_rq = createdb(async_test(), 'database_name', 13);

    open_rq.onupgradeneeded = function(e) {};
    open_rq.onsuccess = function(e) {
        var db = e.target.result;
        assert_equals(db.name, 'database_name', 'db.name');
        assert_equals(db.version, 13, 'db.version');
        this.done();
    }
</script>

<div id=log></div>
